+++ src/js/components/internal/slider-preload.js
@@ -5,4 +5,10 @@ export default {
         target: ({ slides }) => slides,
         targets: (instance) => instance.getAdjacentSlides(),
     }),
+
+    methods: {
+        getAdjacentSlides() {
+            return [1, -1].map((i) => this.slides[this.getIndex(this.index + i)]);
+        },
+    },
 };