+++ src/js/components/slideshow.js
@@ -1,25 +1,24 @@
-import { boxModelAdjust, css, inBrowser } from 'uikit-util';
+import { css } from 'uikit-util';
 import Class from '../mixin/class';
+import SliderParallax from '../mixin/slider-parallax';
 import SliderReactive from '../mixin/slider-reactive';
 import Slideshow from '../mixin/slideshow';
 import SliderPreload from './internal/slider-preload';
 import Animations from './internal/slideshow-animations';
 
-const supportsAspectRatio = inBrowser && CSS.supports('aspect-ratio', '1/1');
-
 export default {
-    mixins: [Class, Slideshow, SliderReactive, SliderPreload],
+    mixins: [Class, Slideshow, SliderReactive, SliderParallax, SliderPreload],
 
     props: {
         ratio: String,
-        minHeight: Number,
-        maxHeight: Number,
+        minHeight: String,
+        maxHeight: String,
     },
 
     data: {
         ratio: '16:9',
-        minHeight: false,
-        maxHeight: false,
+        minHeight: undefined,
+        maxHeight: undefined,
         selList: '.uk-slideshow-items',
         attrItem: 'uk-slideshow-item',
         selNav: '.uk-slideshow-nav',
@@ -28,45 +27,14 @@ export default {
 
     watch: {
         list(list) {
-            if (list && supportsAspectRatio) {
-                css(list, {
-                    aspectRatio: this.ratio.replace(':', '/'),
-                    minHeight: this.minHeight || '',
-                    maxHeight: this.maxHeight || '',
-                    minWidth: '100%',
-                    maxWidth: '100%',
-                });
-            }
+            css(list, {
+                aspectRatio: this.ratio ? this.ratio.replace(':', '/') : undefined,
+                minHeight: this.minHeight,
+                maxHeight: this.maxHeight,
+                minWidth: '100%',
+                maxWidth: '100%',
+            });
         },
-    },
-
-    update: {
-        // deprecated: Remove with iOS 17
-        read() {
-            if (!this.list || supportsAspectRatio) {
-                return false;
-            }
-
-            let [width, height] = this.ratio.split(':').map(Number);
-
-            height = (height * this.list.offsetWidth) / width || 0;
-
-            if (this.minHeight) {
-                height = Math.max(this.minHeight, height);
-            }
-
-            if (this.maxHeight) {
-                height = Math.min(this.maxHeight, height);
-            }
-
-            return { height: height - boxModelAdjust(this.list, 'height', 'content-box') };
-        },
-
-        write({ height }) {
-            height > 0 && css(this.list, 'minHeight', height);
-        },
-
-        events: ['resize'],
     },
 
     methods: {