+++ src/js/components/tooltip.js
@@ -169,13 +169,7 @@ function getAlignment(el, target, [dir, align]) {
     }
 
     const props = includes(properties[0], dir) ? properties[1] : properties[0];
-    if (elOffset[props[0]] === targetOffset[props[0]]) {
-        align = props[0];
-    } else if (elOffset[props[1]] === targetOffset[props[1]]) {
-        align = props[1];
-    } else {
-        align = 'center';
-    }
+    align = props.find((prop) => elOffset[prop] === targetOffset[prop]) || 'center';
 
     return [dir, align];
 }