+++ src/js/util/selector.js
@@ -64,7 +64,9 @@ function _query(selector, context = document, queryFn) {
             }
         }
 
-        context = document;
+        if (!isDocument(context)) {
+            context = context.ownerDocument;
+        }
     }
 
     try {
@@ -74,7 +76,7 @@ function _query(selector, context = document, queryFn) {
     }
 }
 
-const selectorRe = /.*?[^\\](?:,|$)/g;
+const selectorRe = /.*?[^\\](?![^(]*\))(?:,|$)/g;
 
 const splitSelector = memoize((selector) =>
     selector.match(selectorRe).map((selector) => selector.replace(/,$/, '').trim()),