+++ src/js/util/viewport.js
@@ -1,4 +1,5 @@
-import { offset, offsetPosition } from './dimensions';
+import { hasClass } from './class';
+import { dimensions, offset, offsetPosition } from './dimensions';
 import { isVisible, parent, parents } from './filter';
 import {
     clamp,
@@ -128,7 +129,7 @@ export function scrolledOver(element, startOffset = 0, endOffset = 0) {
     const start = Math.max(0, elementOffsetTop - viewportHeight + startOffset);
     const end = Math.min(maxScroll, elementOffsetTop + element.offsetHeight - endOffset);
 
-    return clamp((scrollTop - start) / (end - start));
+    return start < end ? clamp((scrollTop - start) / (end - start)) : 1;
 }
 
 export function scrollParents(element, scrollable = false, props = []) {
@@ -200,9 +201,12 @@ export function offsetViewport(scrollElement) {
 }
 
 export function getCoveringElement(target) {
-    return target.ownerDocument.elementsFromPoint(offset(target).left, 0).find(
+    const { left, width, top } = dimensions(target);
+    return target.ownerDocument.elementsFromPoint(left + width / 2, top).find(
         (el) =>
             !el.contains(target) &&
+            // If e.g. Offcanvas is not yet closed
+            !hasClass(el, 'uk-togglable-leave') &&
             ((hasPosition(el, 'fixed') &&
                 zIndex(
                     parents(target)